rename cleo, coords, viz
