+++ CONTRIBUTING.md
@@ -57,6 +57,6 @@ your IDE with auto-formatting on save.
 I was going to lint using flake8 but then I realized, this is a small research code package! We don't need super pretty, consistent code. Just try to follow Python conventions and use Black.
 
 ## Structure
-Originally, the intention was for opto and electrodes to live under stimulators and recorders, respectively. This made `opto_stim = cleosim.opto.OptogeneticIntervention(...)` possible but not for importing from that second-level shortcut (`from cleosim.opto import ...`). Thus, they were moved up a level. 
+Originally, the intention was for opto and electrodes to live under stimulators and recorders, respectively. This made `opto_stim = cleo.opto.OptogeneticIntervention(...)` possible but not for importing from that second-level shortcut (`from cleo.opto import ...`). Thus, they were moved up a level. 
 
-We still have some import shortcuts for users, making everything in the `electrodes` subpackage (the contents of lfp, spiking, and probes modules) available under `cleosim.electrodes`. We do this by importing the submodules' contents in `__init__.py` files. We can then test the shorcut imports by making sure to use them in the unit tests. However, we must use the full import path in the source code itself to avoid circular import errors. 
+We still have some import shortcuts for users, making everything in the `electrodes` subpackage (the contents of lfp, spiking, and probes modules) available under `cleo.electrodes`. We do this by importing the submodules' contents in `__init__.py` files. We can then test the shorcut imports by making sure to use them in the unit tests. However, we must use the full import path in the source code itself to avoid circular import errors.
