+++ cleo/__init__.py
@@ -1,0 +1,17 @@
+# auto-import submodules
+import cleo.electrodes
+import cleo.opto
+import cleo.coords
+import cleo.stimulators
+import cleo.recorders
+import cleo.processing
+import cleo.utilities
+import cleo.viz
+
+from cleo.base import (
+    CLSimulator,
+    Recorder,
+    Stimulator,
+    InterfaceDevice,
+    IOProcessor,
+)
