+++ cleo/electrodes/__init__.py
@@ -1,0 +1,13 @@
+"""Contains probes, coordinate convenience functions, signals, spiking, and LFP"""
+from cleo.electrodes.lfp import TKLFPSignal
+from cleo.electrodes.spiking import MultiUnitSpiking, SortedSpiking, Spiking
+from cleo.electrodes.probes import (
+    Probe,
+    Signal,
+    concat_coords,
+    linear_shank_coords,
+    tetrode_shank_coords,
+    poly2_shank_coords,
+    poly3_shank_coords,
+    tile_coords,
+)
