+++ cleo/processing/__init__.py
@@ -1,0 +1,9 @@
+"""Classes and functions for constructing and configuring an :class:`~cleo.IOProcessor`."""
+from cleo.processing.base import (
+    LatencyIOProcessor,
+    ProcessingBlock,
+    RecordOnlyProcessor,
+)
+from cleo.processing.controllers import PIController
+from cleo.processing.observers import FiringRateEstimator
+from cleo.processing.delays import Delay, ConstantDelay, GaussianDelay
