--- cleosim/__init__.py
@@ -1,17 +1,0 @@
-# auto-import submodules
-import cleosim.electrodes
-import cleosim.opto
-import cleosim.coordinates
-import cleosim.stimulators
-import cleosim.recorders
-import cleosim.processing
-import cleosim.utilities
-import cleosim.visualization
-
-from cleosim.base import (
-    CLSimulator,
-    Recorder,
-    Stimulator,
-    InterfaceDevice,
-    IOProcessor,
-)
