--- cleosim/processing/__init__.py
@@ -1,9 +1,0 @@
-"""Classes and functions for constructing and configuring an :class:`~cleosim.IOProcessor`."""
-from cleosim.processing.base import (
-    LatencyIOProcessor,
-    ProcessingBlock,
-    RecordOnlyProcessor,
-)
-from cleosim.processing.controllers import PIController
-from cleosim.processing.observers import FiringRateEstimator
-from cleosim.processing.delays import Delay, ConstantDelay, GaussianDelay
