+++ docs/index.rst
@@ -1,4 +1,4 @@
-.. cleosim documentation master file, created by
+.. cleo documentation master file, created by
    sphinx-quickstart on Wed Jan 26 10:53:24 2022.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
