+++ tests/test_coordinates.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from brian2 import NeuronGroup, mm
 
-from cleosim.coordinates import *
+from cleo.coords import *
 
 
 def test_rect_prism_grid():
@@ -29,15 +29,19 @@ def test_cylinder_random():
     ng = NeuronGroup(100, "v=-70*mV : volt")
     assign_coords_rand_cylinder(ng, (1, 1, 1), (2, 2, 2), 1)
     # none past the ends
-    assert not np.any(np.logical_and.reduce((ng.x < 1 * mm, ng.y < 1 * mm, ng.z < 1 * mm)))
-    assert not np.any(np.logical_and.reduce((ng.x > 2 * mm, ng.y > 2 * mm, ng.z > 2 * mm)))
+    assert not np.any(
+        np.logical_and.reduce((ng.x < 1 * mm, ng.y < 1 * mm, ng.z < 1 * mm))
+    )
+    assert not np.any(
+        np.logical_and.reduce((ng.x > 2 * mm, ng.y > 2 * mm, ng.z > 2 * mm))
+    )
     # none exactly on the axis (theoretically possible but highly improbable)
     assert not np.any(ng.z == ng.x / 2 + ng.y / 2)
 
     assign_coords_rand_cylinder(ng, (0, 0, 0), (0, 0, 1), 1)
     # none past the ends
-    assert np.all(ng.z <= 1*mm)
-    assert np.all(ng.z >= 0*mm)
+    assert np.all(ng.z <= 1 * mm)
+    assert np.all(ng.z >= 0 * mm)
     # none exactly on the axis (theoretically possible but highly improbable)
     assert not np.any(ng.z == ng.x / 2 + ng.y / 2)
 
@@ -46,15 +50,19 @@ def test_cylinder_uniform():
     ng = NeuronGroup(100, "v=-70*mV : volt")
     assign_coords_uniform_cylinder(ng, (1, 1, 1), (2, 2, 2), 1)
     # none past the ends
-    assert not np.any(np.logical_and.reduce((ng.x < 1 * mm, ng.y < 1 * mm, ng.z < 1 * mm)))
-    assert not np.any(np.logical_and.reduce((ng.x > 2 * mm, ng.y > 2 * mm, ng.z > 2 * mm)))
+    assert not np.any(
+        np.logical_and.reduce((ng.x < 1 * mm, ng.y < 1 * mm, ng.z < 1 * mm))
+    )
+    assert not np.any(
+        np.logical_and.reduce((ng.x > 2 * mm, ng.y > 2 * mm, ng.z > 2 * mm))
+    )
     # none exactly on the axis (theoretically possible but highly improbable)
     assert not np.any(ng.z == ng.x / 2 + ng.y / 2)
 
     assign_coords_rand_cylinder(ng, (0, 0, 0), (0, 0, 1), 1)
     # none past the ends
-    assert np.all(ng.z <= 1*mm)
-    assert np.all(ng.z >= 0*mm)
+    assert np.all(ng.z <= 1 * mm)
+    assert np.all(ng.z >= 0 * mm)
     # none exactly on the axis (theoretically possible but highly improbable)
     assert not np.any(ng.z == ng.x / 2 + ng.y / 2)
